recruitment-ai/requirements.txt
numpy==1.26.4
pydantic==1.10.13
langchain==0.1.12
langchain-community==0.0.29
chromadb==0.4.22
sentence-transformers==2.2.2
python-dotenv
requests==2.31.0
python-dateutil==2.8.2
colorlog==6.7.0
python-docx>=1.1.0
PyPDF2==3.0.1
pypdf>=3.15.1,<4.0.0
unstructured[docx,pdf]>=0.10.30,<0.11.0
twilio==8.5.0
ollama>=0.1.0
langsmith>=0.1.0,<0.2.0